Bugfix: `locals` should consolidate redundant DIEs for variables/args.
A single formal arg or local var in a method may have multiple
FormalArg or VariableInfo entries within the DW_TAG_subprogram or
redefined within a DW_TAG_lexical_block, offering the same or different
PC-specific instructions for computing the location of the value. Do not
print these on separate, repeat lines -- instead, consolidate the
entries and report the best result (the successfully-located result with
the fewest warnings, or if multiple clean results, the one computed with
the location-based instructions closest to the $PC for the breakpoint.